# src/main/java/com/example/bankrest/entity/CardOperationHistory.java
package com.example.bankrest.entity;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Entity
@Table(name = "card_operations_history")
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CardOperationHistory {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "card_id", nullable = false)
    private Card card;

    @Enumerated(EnumType.STRING)
    @Column(name = "operation_type", nullable = false)
    private OperationType operationType;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "performed_by_id", nullable = false)
    private User performedBy;

    @Enumerated(EnumType.STRING)
    @Column(name = "previous_status")
    private CardStatus previousStatus;

    @Enumerated(EnumType.STRING)
    @Column(name = "new_status")
    private CardStatus newStatus;

    @Column(name = "comment")
    private String comment;

    @Column(name = "created_at", nullable = false)
    private LocalDateTime createdAt;

    public enum OperationType {
        BLOCK, ACTIVATE, DELETE, CREATE
    }
} 